Resolve the `mix deps.get` concurrency issue
If multiple instances of this run in parallel, one of them will fail in
various non-obvious ways. Since resolving deps now has a common series
of steps, Dagger is able to de-duplicate the ops and run it just once,
before the two pipelines diverge.

Running this in the new inline TUI - `export
_EXPERIMENTAL_DAGGER_TUI=inline` - makes this easy to understand.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>