Replace Dagger in Docker with Dagger running directly on Fly.io
There is a new GHA workflow - dagger_on_fly.yml - that makes use of this
new Engine. There are also two new mage targets that manages the remote
Dagger Engine. We use them to start the remote Dagger Engine in the
pipeline, on-demand, and stop it when the job finishes running. If the
job fails before it attempts to stop the Engine, this will remain
running. This is exactly what we want since re-running failed jobs
should be quicker (only by 20s, Dagger Engines are quick to start on
Fly.io).

The other interesting aspect of this is that we have a primary Engine,
and a secondary one in case the primary one fails (regardless of the
reason). Yes, always run 2 of everything™.

Fly-by improvement: bump direnv to latest.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>