Add status.changelog.com (#470)
+ add link in the footer
+ add to our infrastructure diagram

You may remember us mentioning in
https://changelog.com/friends/2#transcript-369 that we switched off
Pingdom. As it turns out, we still want external, synthetic monitoring
that offers a public status page that we can use to communicate
incidents with the outside world. BetterStack.com seems to fit nicely.
Check it out: https://status.changelog.com

We are currently on the Basic (free) tier. This limits us to a 3-minute
check interval & 10 monitors, which is good-enough for now. Leaving this
one with Adam ; ).

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>