Run Dagger Engine as a Fly Machine (no more Docker) (#471)
* Add instructions on how to provision a Dagger Engine on Fly.io

+ includes `mage ci` benchmarks for various VM sizes
+ stress-tested Fly.io apps v2 by creating volumes & machines at least
  30 times across several days (all worked as expected!)

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Share the same Dagger Client across all mage tasks

Rather than opening multiple connections from each task when they run in
parallel, open one connection and share it via Go's context.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Bump Node.js to latest v18

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Resolve the `mix deps.get` concurrency issue

If multiple instances of this run in parallel, one of them will fail in
various non-obvious ways. Since resolving deps now has a common series
of steps, Dagger is able to de-duplicate the ops and run it just once,
before the two pipelines diverge.

Running this in the new inline TUI - `export
_EXPERIMENTAL_DAGGER_TUI=inline` - makes this easy to understand.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Replace Dagger in Docker with Dagger running directly on Fly.io

There is a new GHA workflow - dagger_on_fly.yml - that makes use of this
new Engine. There are also two new mage targets that manages the remote
Dagger Engine. We use them to start the remote Dagger Engine in the
pipeline, on-demand, and stop it when the job finishes running. If the
job fails before it attempts to stop the Engine, this will remain
running. This is exactly what we want since re-running failed jobs
should be quicker (only by 20s, Dagger Engines are quick to start on
Fly.io).

The other interesting aspect of this is that we have a primary Engine,
and a secondary one in case the primary one fails (regardless of the
reason). Yes, always run 2 of everything™.

Fly-by improvement: bump direnv to latest.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Fix OCI image labels

https://github.com/dagger/dagger/issues/5535

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

---------

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>