Bump deps to latest
This started as a flyctl upgrade (the version that we were using was
unsupported), and then moved to everything else.

It's been a while (>4 months), time to pull the latest minors and see if anything
breaks:
- Ubuntu 22.04 LTS (Jammy Jellyfish)
- Elixir 1.14.5
- Golang 1.20.12
- flyctl 0.1.134

For Node.js, pull the latest Active LTS, which is currently:
- Node.js 20.10.0
FTR: https://nodejs.org/en/about/previous-releases

For Erlang, we have a few minors in 26 already, and everything works as
expected, so let's bump to latest stable:
- Erlang 26.2

We should also bump Elixir to latest minor (currently 1.15) as a
follow-up. There were issues with it, will add more details to the PR.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>