Upload static assets when app boots
Since uploading static assets requires a db connection - doesn't seem
right, but it is what it is - and prod config requires Postgres to use
SSL, and we should not connect to the production db from the CI, the
only option is to upload these on app boot, just before we run db
migrations.

If uploading static assets would not require a db connection, this would
definitely be a CI concern, but for now this is the simplest fix.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>