Enable changelog.com devs to create prod db forks with a single command
This is the next logical step after migrating to Neon.tech part of
https://github.com/thechangelog/changelog.com/pull/492

    cd changelog
    dagger call db-branch --neon-api-key=env:NEON_API_KEY

To learn more, see `changelog/README.md`. Part of this, we also deployed
a Dagger Engine v0.10.3 on Fly.io so that we don't need any sort of
container runtime running locally. I know that Jerod will appreciate
this.

The beginning of a new generation of tooling, I'm sure of it.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>