Enable changelog.com devs to create prod db forks with a single command (#508)
* Downgrade Erlang to 26.2.2

Elixir 1.14.5 installed was segfaulting on macOS 12.7.3 ARM. Installing
the Elixir `otp26` variant didn't fix it. Maybe an asdf issue... Will
try again another time.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Enable changelog.com devs to create prod db forks with a single command

This is the next logical step after migrating to Neon.tech part of
https://github.com/thechangelog/changelog.com/pull/492

    cd changelog
    dagger call db-branch --neon-api-key=env:NEON_API_KEY

To learn more, see `changelog/README.md`. Part of this, we also deployed
a Dagger Engine v0.10.3 on Fly.io so that we don't need any sort of
container runtime running locally. I know that Jerod will appreciate
this.

The beginning of a new generation of tooling, I'm sure of it.

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

---------

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>