Make the time-to-deploy 2x quicker: 2mins vs 4mins
This replaces our Dagger on Fly.io setup with Dagger on Namespace.so

Before this change, throughout August 2024:
- deploys took `4m 36s` 95% of the time
- `1` in `14` deploys failed

Let's see how the above changes after a few weeks of running.

This is the first part of the required change. The second part will
follow after this gets merged. We need to:
- Cleanup secrets & variables from GitHub Actions
- De-provision the Dagger Engine on Fly.io
- Remove the deploy config from fly.io dir
- Update the INFRASTRUCTURE diagram

We still keep K8s to see if we can improve on Namespace.so, the new
Speed King 🏎️

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>