Enable team members to replace changelog_dev with a prod db dump (#533)
* Enable team members to replace changelog_dev with a prod db dump

There is a single command to run:

    just restore-dev-db-from-prod

This implies that:
- direnv.net is hooked into the shell
-`just envrc-secrets` has already run & `.envrc.secrets` exists
- Posgres 16 has been installed using `just postgres`

Part of this change we exposed a few more private recipes and made it
clear who they are meant for:
- `contributor` means anyone that wants to play with this repo locally
- `team` means Adam, Jerod & Gerhard

We also added `just add-oban-pro-repo` so that we make this integration
easier for team members - I always forget about it!

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

* Go back to managing Postgres with asdf

Single tool to rule them all ✊ icu4c seems to be working on macOS 14.7
ARM now. Checked that the new command works as expected. Also confirmed
that CI works locally.

Also run:

    npx browserslist@latest --update-db

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>

---------

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>