Promote changelog-2025-05-05.fly.dev to new production
The new setup has been serving production traffic for a few months now.
Given all the metrics, we are confident that it's time to switch
everything across.

Since 98% of the requests are cached in the new setup, we can
comfortably reduce the app size by 50%, and still deliver much better
performance than before. Stay tuned, we'll unpack this in Kaizen 21:
https://github.com/thechangelog/changelog.com/discussions/551

May the Pipedream be with you!

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>